# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Elite Tenancy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /matching/score:
    post:
      operationId: getMatchingScore
      tags: [matching]
      summary: AI-powered tenant-to-listing match scoring
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantMatchRequest"
      responses:
        "200":
          description: Ranked list of matching properties with AI scores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantMatchResponse"
...
  /tenant/maintenance:
    get:
      operationId: getTenantMaintenance
      tags: [tenant]
      summary: Get tenant's maintenance requests
      responses:
        "200":
          description: Maintenance requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceRequest"
    post:
      operationId: submitMaintenanceRequest
      tags: [tenant]
      summary: Submit a maintenance request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceInput"
      responses:
        "201":
          description: Maintenance request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"
...
components:
  schemas:
...
    Listing:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        price: { type: integer }
        pricePeriod: { type: string }
        category: { type: string }
        city: { type: string }
        postcode: { type: string }
        addressLine1: { type: string }
        bedrooms: { type: integer }
        bathrooms: { type: integer }
        floorAreaSqm: { type: ["integer", "null"] }
        furnished: { type: boolean }
        petsAllowed: { type: boolean }
        billsIncluded: { type: boolean }
        status: { type: string }
        isFeatured: { type: boolean }
        isPremium: { type: boolean }
        photos: { type: array, items: { type: string } }
        aiMatchScore: { type: ["integer", "null"] }
        availableFrom: { type: ["string", "null"] }
        viewCount: { type: integer }
        enquiryCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        - id
        - title
        - slug
        - description
        - price
        - pricePeriod
        - category
        - city
        - postcode
        - bedrooms
        - bathrooms
        - furnished
        - petsAllowed
        - billsIncluded
        - status
        - isFeatured
        - isPremium
        - photos
        - viewCount
        - enquiryCount
        - createdAt
        - updatedAt
...
    MaintenanceRequest:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        priority: { type: string }
        status: { type: string }
        propertyAddress: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required:
        - id
        - title
        - description
        - category
        - priority
        - status
        - createdAt
        - updatedAt
    MaintenanceInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        priority: { type: string }
      required:
        - title
        - description
        - category
        - priority
...
    TenantMatchRequest:
      type: object
      properties:
        city: { type: string }
        minBudget: { type: number }
        maxBudget: { type: number }
        bedrooms: { type: number }
        furnished: { type: boolean }
        petsAllowed: { type: boolean }
        billsIncluded: { type: boolean }
        moveInDate: { type: string }
        lifestyle: { type: string }
        priorities: { type: string }
      required:
        - maxBudget
    TenantMatchResult:
      type: object
      properties:
        id: { type: integer }
        score: { type: integer }
        summary: { type: string }
        highlights:
          type: array
          items: { type: string }
        listing:
          $ref: "#/components/schemas/Listing"
      required:
        - id
        - score
        - summary
        - highlights
        - listing
    TenantMatchResponse:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: "#/components/schemas/TenantMatchResult"
      required:
        - matches