# lib/api-zod/src/generated/types/maintenanceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Elite Tenancy API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MaintenanceInput {
  title: string;
  description: string;
  category: string;
  priority: string;
}